+++ src/collective/nitf/browser/__init__.py
@@ -3,7 +3,7 @@ from plone import api
 from plone.app.imaging.scaling import ImageScaling as BaseImageScaling
 from plone.app.layout.viewlets.content import DocumentBylineViewlet
 from plone.dexterity.browser.view import DefaultView
-from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from plone.memoize import view
 
 
 class View(DefaultView):
@@ -66,22 +66,37 @@ class ImageScaling(BaseImageScaling):
 
 class NITFBylineViewlet(DocumentBylineViewlet):
 
-    index = ViewPageTemplateFile('templates/nitf_byline.pt')
+    """Override the document byline viewlet to include semantic markup."""
 
-    def getMemberInfoByName(self, fullname):
+    @view.memoize
+    def _search_member_by_name(self, fullname):
+        """Search a user by its full name and return its member
+        information.
+
+        :param fullname: full name of the user we are looking for
+        :type fullname: unicode
+        :returns: member information
+        :rtype: dict
+        """
+        if not fullname:
+            return None
+
         membership = api.portal.get_tool('portal_membership')
         members = membership.searchForMembers(name=fullname)
         if members:
-            member = members[0].getUserId()  # we care only about the first
+            # in case there are more than one members with the
+            # same fullname, we use the first one listed
+            member = members[0].getUserId()
             return membership.getMemberInfo(member)
 
-    def byline(self):
-        member = self.getMemberInfoByName(self.context.byline)
+    @property
+    def author_id(self):
+        member = self._search_member_by_name(self.context.byline)
         if member:
             return member['username']
 
     def author(self):
-        return self.getMemberInfoByName(self.context.byline)
+        return self._search_member_by_name(self.context.byline)
 
     def authorname(self):
         return self.context.byline