+++ CHANGES.rst
@@ -9,6 +9,9 @@ There's a frood who really knows where his towel is.
 .. Warning::
     Upgrades are supported only from release 1.0b3.
 
+- Replace Galleria with Cycle2 (closes `#116`_).
+  [rodfersou][hvelarde]
+
 - Fix default values for genre and urgency fields (closes `#118`_).
   [rodfersou]
 
@@ -248,5 +251,6 @@ There's a frood who really knows where his towel is.
 .. _`#94`: https://github.com/collective/collective.nitf/issues/94
 .. _`#98`: https://github.com/collective/collective.nitf/issues/98
 .. _`#111`: https://github.com/collective/collective.nitf/issues/111
+.. _`#116`: https://github.com/collective/collective.nitf/issues/116
 .. _`#118`: https://github.com/collective/collective.nitf/issues/118
 .. _`#123`: https://github.com/collective/collective.nitf/issues/123