+++ CHANGES.rst
@@ -9,11 +9,10 @@ There's a frood who really knows where his towel is.
 .. Warning::
     Upgrades are supported only from release 1.0b3.
 
-- Replace Galleria with Cycle2 (closes `#116`_).
-  [rodfersou][hvelarde]
+- Replace slideshow framework machinery; we use `Cycle2`_ now instead of `Galleria`_ (closes `#116`_).
+  [rodfersou, hvelarde]
 
 - Fix default values for genre and urgency fields (closes `#118`_).
-  [rodfersou]
 
 - Add NITF tile for collective.cover (closes `#123`_).
   [hvelarde]
@@ -238,6 +237,8 @@ There's a frood who really knows where his towel is.
 
 - Initial release.
 
+.. _`Cycle2`: http://jquery.malsup.com/cycle2/
+.. _`Galleria`: http://galleria.io/
 .. _`#33`: https://github.com/collective/collective.nitf/issues/33
 .. _`#47`: https://github.com/collective/collective.nitf/issues/47
 .. _`#49`: https://github.com/collective/collective.nitf/issues/49