+++ CHANGES.rst
@@ -9,7 +9,8 @@ There's a frood who really knows where his towel is.
 .. Warning::
     Upgrades are supported only from release 1.0b3.
 
-- Fix display of byline in ``documentbyline`` viewlet (fixes `#100`_).
+- Fix display of byline and refactor override of ``documentbyline`` viewlet;
+  avoid performance issues when having many users by memoizing expensive call to Membership tool (fixes `#128`_).
   [hvelarde, rodfersou]
 
 - Replace slideshow framework machinery;
@@ -234,8 +235,8 @@ There's a frood who really knows where his towel is.
 .. _`#93`: https://github.com/collective/collective.nitf/issues/93
 .. _`#94`: https://github.com/collective/collective.nitf/issues/94
 .. _`#98`: https://github.com/collective/collective.nitf/issues/98
-.. _`#100`: https://github.com/collective/collective.nitf/issues/100
 .. _`#111`: https://github.com/collective/collective.nitf/issues/111
 .. _`#116`: https://github.com/collective/collective.nitf/issues/116
 .. _`#118`: https://github.com/collective/collective.nitf/issues/118
 .. _`#123`: https://github.com/collective/collective.nitf/issues/123
+.. _`#128`: https://github.com/collective/collective.nitf/issues/128